重构代码，优化语言处理函数并注册SQL语言支持
