+++ src/utils/hljs.ts
@@ -4,6 +4,7 @@ import java from 'highlight.js/lib/languages/java'
 import javascript from 'highlight.js/lib/languages/javascript'
 import python from 'highlight.js/lib/languages/python'
 import xml from 'highlight.js/lib/languages/xml'
+import sql from 'highlight.js/lib/languages/sql'
 
 // Legacy markdown uses ```in / ```out fences for sample IO; register no-op languages to avoid warnings.
 hljs.registerLanguage('in', () => ({ name: 'in', contains: [] }))
@@ -14,5 +15,59 @@ hljs.registerLanguage('javascript', javascript)
 hljs.registerLanguage('css', css)
 hljs.registerLanguage('xml', xml)
 hljs.registerLanguage('html', xml)
+hljs.registerLanguage('sql', sql)
 
-export default hljs
+function normalizeLanguage(language?: string) {
+  if (!language) return ''
+  const normalized = language.trim().toLowerCase()
+  if (normalized === 'js') return 'javascript'
+  if (normalized === 'py') return 'python'
+  if (normalized === 'htm') return 'html'
+  return normalized
+}
+
+function getLanguageExtension(language?: string) {
+  switch (normalizeLanguage(language)) {
+    case 'java':
+      return 'java'
+    case 'python':
+      return 'py'
+    case 'javascript':
+      return 'js'
+    case 'html':
+      return 'html'
+    case 'css':
+      return 'css'
+    case 'c':
+      return 'c'
+    default:
+      return 'txt'
+  }
+}
+
+function formatFilename(filename: string | undefined, language?: string) {
+  const trimmed = filename?.trim()
+  if (trimmed) return trimmed
+  return `main.${getLanguageExtension(language)}`
+}
+
+function formatLanguageLabel(language: string | undefined) {
+  const normalized = normalizeLanguage(language)
+  if (!normalized) return 'text'
+  if (normalized === 'html') return 'HTML'
+  if (normalized === 'css') return 'CSS'
+  if (normalized === 'javascript') return 'JavaScript'
+  if (normalized === 'python') return 'Python'
+  if (normalized === 'java') return 'Java'
+  return normalized
+}
+
+function highlightCode(content: string, language?: string) {
+  const normalized = normalizeLanguage(language)
+  if (normalized && hljs.getLanguage(normalized)) {
+    return hljs.highlight(content, { language: normalized }).value
+  }
+  return hljs.highlightAuto(content).value
+}
+
+export { hljs, formatFilename, formatLanguageLabel, highlightCode }
