+++ src/views/CodeView.vue
@@ -34,7 +34,7 @@
 
 <script setup lang="ts">
 import MarkdownIt from 'markdown-it'
-import hljs from '../utils/hljs'
+import { hljs, formatFilename, formatLanguageLabel, highlightCode } from '../utils/hljs'
 // @ts-expect-error No types for markdown-it-katex
 import markdownItKatex from 'markdown-it-katex'
 import MarkdownItHighlight from 'markdown-it-highlightjs'
@@ -178,59 +178,4 @@ async function copyCode(file: { rawCode: string; copied: boolean }) {
     console.error('Failed to copy text: ', err)
   }
 }
-
-function normalizeLanguage(language?: string) {
-  if (!language) return ''
-  const normalized = language.trim().toLowerCase()
-  if (normalized === 'js') return 'javascript'
-  if (normalized === 'py') return 'python'
-  if (normalized === 'htm') return 'html'
-  return normalized
-}
-
-function getLanguageExtension(language?: string) {
-  switch (normalizeLanguage(language)) {
-    case 'java':
-      return 'java'
-    case 'python':
-      return 'py'
-    case 'javascript':
-      return 'js'
-    case 'html':
-      return 'html'
-    case 'css':
-      return 'css'
-    case 'c':
-      return 'c'
-    default:
-      return 'txt'
-  }
-}
-
-function formatFilename(filename: string | undefined, language?: string) {
-  const trimmed = filename?.trim()
-  if (trimmed) return trimmed
-  return `main.${getLanguageExtension(language)}`
-}
-
-function formatLanguageLabel(language: string | undefined) {
-  const normalized = normalizeLanguage(language)
-  if (!normalized) return 'text'
-  if (normalized === 'html') return 'HTML'
-  if (normalized === 'css') return 'CSS'
-  if (normalized === 'javascript') return 'JavaScript'
-  if (normalized === 'python') return 'Python'
-  if (normalized === 'java') return 'Java'
-  return normalized
-}
-
-function highlightCode(content: string, language?: string) {
-  const normalized = normalizeLanguage(language)
-  if (normalized && hljs.getLanguage(normalized)) {
-    return hljs.highlight(content, { language: normalized }).value
-  }
-  return hljs.highlightAuto(content).value
-}
-
-console.log('table_open:', mdParser.renderer.rules.table_open)
 </script>
